unittests/requirements.txt
aws-cdk-lib>=2.38.0
constructs>=10.0.0,<11.0.0
pytest>=6.2.4
boto3>=1.34.77
moto>=5.0.4